Refactored get view claim
